Move plugin API packages into legacy.
