+++ politeiawww/legacy/pi/events.go
@@ -18,10 +18,10 @@ import (
 	tkv1 "github.com/decred/politeia/politeiawww/api/ticketvote/v1"
 	www "github.com/decred/politeia/politeiawww/api/www/v1"
 	"github.com/decred/politeia/politeiawww/client"
-	"github.com/decred/politeia/politeiawww/comments"
+	"github.com/decred/politeia/politeiawww/legacy/comments"
+	"github.com/decred/politeia/politeiawww/legacy/ticketvote"
 	"github.com/decred/politeia/politeiawww/legacy/user"
 	"github.com/decred/politeia/politeiawww/records"
-	"github.com/decred/politeia/politeiawww/ticketvote"
 	"github.com/google/uuid"
 )
