+++ politeiawww/legacy/politeiawww.go
@@ -25,7 +25,6 @@ import (
 	piplugin "github.com/decred/politeia/politeiad/plugins/pi"
 	tkplugin "github.com/decred/politeia/politeiad/plugins/ticketvote"
 	umplugin "github.com/decred/politeia/politeiad/plugins/usermd"
-	"github.com/decred/politeia/politeiawww/comments"
 	"github.com/decred/politeia/politeiawww/config"
 	"github.com/decred/politeia/politeiawww/events"
 	"github.com/decred/politeia/politeiawww/legacy/cmsdatabase"
@@ -33,15 +32,16 @@ import (
 	cmsdb "github.com/decred/politeia/politeiawww/legacy/cmsdatabase/cockroachdb"
 	"github.com/decred/politeia/politeiawww/legacy/codetracker"
 	ghtracker "github.com/decred/politeia/politeiawww/legacy/codetracker/github"
+	"github.com/decred/politeia/politeiawww/legacy/comments"
+	"github.com/decred/politeia/politeiawww/legacy/pi"
+	"github.com/decred/politeia/politeiawww/legacy/ticketvote"
 	"github.com/decred/politeia/politeiawww/legacy/user"
 	"github.com/decred/politeia/politeiawww/legacy/user/cockroachdb"
 	"github.com/decred/politeia/politeiawww/legacy/user/localdb"
 	"github.com/decred/politeia/politeiawww/legacy/user/mysql"
 	"github.com/decred/politeia/politeiawww/mail"
-	"github.com/decred/politeia/politeiawww/pi"
 	"github.com/decred/politeia/politeiawww/records"
 	"github.com/decred/politeia/politeiawww/sessions"
-	"github.com/decred/politeia/politeiawww/ticketvote"
 	"github.com/decred/politeia/util"
 	"github.com/decred/politeia/wsdcrdata"
 	"github.com/google/uuid"
