+++ politeiawww/legacy/ticketvote/events.go
