+++ politeiawww/log.go
@@ -9,20 +9,20 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/decred/politeia/politeiawww/comments"
 	"github.com/decred/politeia/politeiawww/events"
 	"github.com/decred/politeia/politeiawww/legacy"
 	cmsdb "github.com/decred/politeia/politeiawww/legacy/cmsdatabase/cockroachdb"
 	"github.com/decred/politeia/politeiawww/legacy/codetracker/github"
 	ghdb "github.com/decred/politeia/politeiawww/legacy/codetracker/github/database/cockroachdb"
+	"github.com/decred/politeia/politeiawww/legacy/comments"
+	"github.com/decred/politeia/politeiawww/legacy/pi"
+	"github.com/decred/politeia/politeiawww/legacy/ticketvote"
 	"github.com/decred/politeia/politeiawww/legacy/user/cockroachdb"
 	"github.com/decred/politeia/politeiawww/legacy/user/localdb"
 	"github.com/decred/politeia/politeiawww/legacy/user/mysql"
 	"github.com/decred/politeia/politeiawww/mail"
-	"github.com/decred/politeia/politeiawww/pi"
 	"github.com/decred/politeia/politeiawww/records"
 	"github.com/decred/politeia/politeiawww/sessions"
-	"github.com/decred/politeia/politeiawww/ticketvote"
 	"github.com/decred/politeia/wsdcrdata"
 	"github.com/decred/slog"
 	"github.com/jrick/logrotate/rotator"
